# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EdgeEver API",
    "version": "0.1.0",
    "description": "REST API for EdgeEver notebooks, memos, resources, API tokens, and Agent integrations."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/exports/markdown": {
      "get": {
        "summary": "Read a page of Markdown export data",
        "description": "Returns active memo details and their resource metadata for browser-side Markdown ZIP generation. Requires read:memos and read:resources scopes.",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of active memos and related resources",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "memos": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MemoDetail"
                      }
                    },
                    "resources": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Resource"
                      }
                    },
                    "totalCount": {
                      "type": "integer"
                    },
                    "nextOffset": {
                      "type": ["integer", "null"]
                    }
                  },
                  "required": ["memos", "resources", "totalCount", "nextOffset"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backups/json": {
      "get": {
        "summary": "Read a page of EdgeEver ZIP recovery data",
        "description": "Returns active memo details, revision history, and resource metadata for browser-side EdgeEver ZIP generation. The resulting archive also contains human-readable Markdown. Trash is excluded.",
        "parameters": [
          { "name": "offset", "in": "query", "schema": { "type": "integer", "minimum": 0, "default": 0 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 50, "default": 25 } }
        ],
        "responses": {
          "200": {
            "description": "A page of logical backup data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "memos": { "type": "array", "items": { "$ref": "#/components/schemas/MemoDetail" } },
                    "resources": { "type": "array", "items": { "$ref": "#/components/schemas/Resource" } },
                    "revisions": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
                    "totalCount": { "type": "integer" },
                    "nextOffset": { "type": ["integer", "null"] }
                  },
                  "required": ["memos", "resources", "revisions", "totalCount", "nextOffset"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/restores/json/notebooks": {
      "post": {
        "summary": "Import a batch of notebooks from an EdgeEver ZIP",
        "description": "Interactive-user-only merge restore. Matching EdgeEver notebook IDs are overwritten.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "notebooks": { "type": "array", "minItems": 1, "maxItems": 100, "items": { "type": "object", "additionalProperties": true } } },
                "required": ["notebooks"]
              }
            }
          }
        },
        "responses": { "200": { "$ref": "#/components/responses/Ok" } }
      }
    },
    "/api/v1/restores/json/memos": {
      "post": {
        "summary": "Import a batch of memos from an EdgeEver ZIP",
        "description": "Interactive-user-only merge restore, including current content and revision history. Matching EdgeEver memo IDs are overwritten.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "memos": { "type": "array", "minItems": 1, "maxItems": 10, "items": { "type": "object", "additionalProperties": true } } },
                "required": ["memos"]
              }
            }
          }
        },
        "responses": { "200": { "$ref": "#/components/responses/Ok" } }
      }
    },
    "/api/v1/restores/json/resources/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "put": {
        "summary": "Import one resource from an EdgeEver ZIP",
        "description": "Interactive-user-only resource import. The multipart metadata field contains the structured EdgeEver ZIP resource record.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "metadata": { "type": "string" },
                  "file": { "type": "string", "format": "binary" }
                },
                "required": ["metadata", "file"]
              }
            }
          }
        },
        "responses": { "200": { "$ref": "#/components/responses/Ok" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "edgeever_session"
      }
    },
...
    "responses": {
      "Ok": {
        "description": "OK",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Ok"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": ["ok"]
      },
...
      "MemoSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "notebookId": {
            "type": "string"
          },
          "title": {
            "type": ["string", "null"]
          },
          "excerpt": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isPinned": {
            "type": "boolean"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "revision": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": ["string", "null"],
            "format": "date-time"
          }
        },
        "required": ["id", "notebookId", "title", "excerpt", "tags", "isPinned", "isArchived", "isDeleted", "revision", "createdAt", "updatedAt", "deletedAt"]
      },
      "MemoDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MemoSummary"
          },
          {
            "type": "object",
            "properties": {
              "contentJson": {
                "$ref": "#/components/schemas/TiptapDoc"
              },
              "contentMarkdown": {
                "type": "string"
              },
              "contentText": {
                "type": "string"
              },
              "contentHash": {
                "type": "string",
                "minLength": 64,
                "maxLength": 64
              },
              "sourceMemoIds": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "mergeSourceCount": {
                "type": "integer"
              },
              "mergedIntoMemoId": {
                "type": ["string", "null"]
              }
            },
            "required": ["contentJson", "contentMarkdown", "contentText", "contentHash", "sourceMemoIds", "mergeSourceCount", "mergedIntoMemoId"]
          }
        ]
      },
...
      "Resource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "memoId": {
            "type": "string"
          },
          "originalMemoId": {
            "type": ["string", "null"]
          },
          "kind": {
            "type": "string",
            "enum": ["image", "attachment"]
          },
          "mimeType": {
            "type": ["string", "null"]
          },
          "filename": {
            "type": ["string", "null"]
          },
          "byteSize": {
            "type": "integer"
          },
          "sha256": {
            "type": ["string", "null"]
          },
          "width": {
            "type": ["integer", "null"]
          },
          "height": {
            "type": ["integer", "null"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "url": {
            "type": "string"
          }
        },
        "required": ["id", "memoId", "originalMemoId", "kind", "mimeType", "filename", "byteSize", "sha256", "width", "height", "createdAt", "updatedAt", "url"]
      },
...
      "TiptapDoc": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "type": {
            "type": "string",
            "const": "doc"
          },
          "content": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "required": ["type", "content"]
      },
...
    }
  }
}